Rework file structure to avoid route shadowing
