+++ front-api/routes/v1/w/[wId]/members/index.ts
@@ -1,10 +1,42 @@
+import { getMembers } from "@app/lib/api/workspace";
+import type {
+  GetWorkspaceMembersResponseBody,
+  UserType,
+} from "@dust-tt/client";
 import { publicApiApp } from "@front-api/middlewares/ctx";
+import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
+import type { HandlerResult } from "@front-api/middlewares/utils";
 
 import emails from "./emails";
 import validate from "./validate";
 
-// Mounted at /api/v1/w/:wId/members.
+// Mounted at /api/v1/w/:wId/members. publicApiAuth is applied by the parent
+// v1 workspace sub-app, so ctx.get("auth") is always available here.
 const app = publicApiApp();
+
+/**
+ * @ignoreswagger
+ * Admin-only endpoint. Undocumented.
+ */
+app.get(
+  "/",
+  ensureIsAdmin(),
+  async (ctx): HandlerResult<GetWorkspaceMembersResponseBody> => {
+    const auth = ctx.get("auth");
+
+    const { members: users } = await getMembers(auth, { activeOnly: true });
+
+    return ctx.json({
+      users: users.map(
+        (user): Pick<UserType, "sId" | "id" | "email"> => ({
+          sId: user.sId,
+          id: user.id,
+          email: user.email,
+        })
+      ),
+    });
+  }
+);
 
 app.route("/emails", emails);
 app.route("/validate", validate);
